user certifies delete choice

+++ backend/locust/cleanup_experiments.py
@@ -10,5 +10,9 @@ expIds = deleteExperiment.getExperimentIds(protocol, host)
 if option > len(expIds):
     option = len(expIds)
 
-for i in range(option):
-        deleteExperiment.deleteExperiment(protocol, host, expIds)
+print(f"There are {len(expIds)} locust test experiments")
+doublecheck = input(f"Are you sure you want to delete {option} of them? Y/N: ")
+
+if doublecheck == "Y" or doublecheck == "y":
+    for i in range(option):
+            deleteExperiment.deleteExperiment(protocol, host, expIds)
